+++ lib/validator/v3/doc.go
@@ -17,4 +17,5 @@ Package validator implements common field and structure validation that is
 used to validate the API structures supplied on the client interface, and
 is also used internally to validate the information stored in the backend
 datastore.
-*/package validator
+*/
+package v3
