+++ lib/validator/v3/validator.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package validator
+package v3
 
 import (
 	"net"
